Fix more spellings of 'canonical'. Move some page versioning functions to other files. Add the page tree fix to the migration that requires the tree to be fixed.
