+++ cms/apps/pages/templates/admin/pages/page/versions_list.html
@@ -37,11 +37,11 @@
     <a href="{% url 'admin:pages_page_changelist' %}">Page</a>
     <span class="icon-arrow-right breadcrumbs-separator"></span>
     {% if display_language_options %}
-      <a href="{{ cannonical_version.get_admin_url }}">{{ cannonical_version }}</a>
+      <a href="{{ canonical_version.get_admin_url }}">{{ canonical_version }}</a>
       <span class="icon-arrow-right breadcrumbs-separator"></span>
       {{ original.get_language }}
     {% else %}
-      {{ cannonical_version }}
+      {{ canonical_version }}
     {% endif %}
   </div>
 
@@ -69,7 +69,7 @@
             </option>
           {% endfor %}
         </select>
-        <a class="button add-new-button" href="{% url 'admin:pages_page_duplicate' object_id=cannonical_version.id %}">Add translation</a>
+        <a class="button add-new-button" href="{% url 'admin:pages_page_duplicate' object_id=canonical_version.id %}">Add translation</a>
       </div>
     {% endif %}
